+++ packages/kitbook/src/lib/routes/(main)/+layout.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import { Layout } from 'kitbook';
+  import Layout from '../../layout/Layout.svelte';
 </script>
 
-<Layout expanded title="Kitbook" githubURL="https://github.com/jacob-8/kitbook/tree/main/packages/kitbook">
+<Layout>
   <slot />
 </Layout>