+++ packages/svelte-pieces/src/.kitbook/WrapRootLayout.svelte
@@ -1,9 +1,18 @@
-<script>import { Layout } from "kitbook";
+<script>
+  import { setContext } from 'svelte';
+
+  /** @type {import('kitbook').KitbookSettings} */
+  const kitbookSettings = {
+    title: 'Svelte Pieces',
+    description: 'Useful Svelte UI pieces for building web apps',
+    githubURL: 'https://github.com/jacob-8/kitbook/tree/main/packages/svelte-pieces',
+    expandTree: true,
+  };
+
+  setContext('kitbook-settings', kitbookSettings);
 </script>
 
-<Layout expanded title="Svelte Pieces" githubURL="https://github.com/jacob-8/kitbook/tree/main/packages/svelte-pieces">
-  <slot />
-</Layout>
+<slot />
 
 <style uno:preflights global>
   [type='text'],