+++ packages/vite-plugin-kitbook/src/constants.ts
@@ -1,2 +1,3 @@
-export const DEFAULT_KITBOOK_ROUTES = 'src/kitbook';
 export const MDSVEX_EXTENSIONS = ['.md', '.svx'];
+export const AUGMENT_FUNCTION_TEXT = `import { augmentSvelteConfigForKitbook } from 'kitbook/plugins/vite'; 
+export default augmentSvelteConfigForKitbook(config)`;